KOKKOS: remove silent fp32/fp64 conversions package-wide
Continue the fp32/fp64 cleanup started in #4754 and extend it to the full set
of packages that provide KOKKOS styles. The conversions were found by building
with clang and -Wall -Wextra -Wimplicit-float-conversion -Wdouble-promotion
-pedantic in both the single- and the mixed-precision configuration, and then
making every flagged conversion explicit. Both builds are now free of
-Wimplicit-float-conversion and -Wdouble-promotion warnings.

The changes follow a few repeated patterns:

- Bare libm calls (sqrt, pow, exp, log, sin, cos, erfc, fabs, ...) on KK_FLOAT
  arguments resolve to the double overload and promote, so they are qualified
  with Kokkos:: to select the float overload. Note that Kokkos::pow() with an
  integer exponent still promotes, so those exponents are cast to KK_FLOAT.
- Double literals and constants (1.0, 0.5, MY_PIS, EWALD_P, A1-A5, MY_EPSILON
  and similar) mixed into KK_FLOAT expressions are cast to KK_FLOAT, which
  folds at compile time.
- Base-class double scalars used inside kernels are copied once into a local
  KK_FLOAT, using the existing _kk suffix convention.
- Values that accumulate into EV_FLOAT fields, the per-atom eatom/vatom views
  or the atomic force and torque views are cast to KK_ACC_FLOAT, while values
  read back out of those views into float math are cast to KK_FLOAT. The
  direction was chosen per site from the declared view type. This distinction
  only matters in the mixed build, which is why it was built and cleaned as a
  separate configuration.
- Host-side reductions into the base-class double eng_vdwl, eng_coul and
  virial[] are cast to double.

Two changes go beyond adding a cast:

pair_table_kokkos compute_fpair(), compute_evdwl() and compute_ecoul() now take
and return KK_FLOAT like every other pair style instead of double. The table
interpolation still runs in double internally through an explicit local and
converts at the return. Previously this style's double interface propagated
into the shared PairComputeFunctor and forced rsq to promote there for every
style. The STACKPARAMS cutoff comparison in pair_kokkos.h is cast to KK_FLOAT
accordingly, which is a no-op for the styles whose cutsq is already KK_FLOAT.

In pair_tip4p_kokkos.h apply_site_force() the O-branch force components fOx and
fHx are declared KK_FLOAT to match fdx in the same function and the sibling
non-O branch above it, so the virial products are formed in KK_FLOAT and widened
once at the accumulator. This is the only change that is not bit-identical: in
the mixed build alone those virial products are formed in float rather than
double. The single and double builds are unaffected.

Everything else is numerically inert: in the default double build every added
cast is a no-op, so results are unchanged.

+++ src/KOKKOS/comm_kokkos.cpp
@@ -996,7 +996,7 @@ struct BuildExchangeListFunctor {
 // NOLINTNEXTLINE
   KOKKOS_INLINE_FUNCTION
   void operator() (int i) const {
-    if (_x(i,_dim) < _lo || _x(i,_dim) >= _hi) {
+    if (static_cast<double>(_x(i,_dim)) < _lo || static_cast<double>(_x(i,_dim)) >= _hi) {
       const int mysend = Kokkos::atomic_fetch_add(&_nsend(0),1);
       if (mysend < (int)_sendlist.extent(0))
         _sendlist(mysend) = i;
@@ -1437,14 +1437,14 @@ struct BuildBorderListFunctor {
     const int teamend = (teamstart + chunk) < nlast?(teamstart + chunk):nlast;
     int mysend = 0;
     for (int i=teamstart + dev.team_rank(); i<teamend; i+=dev.team_size()) {
-      if (x(i,dim) >= lo && x(i,dim) <= hi) mysend++;
+      if (static_cast<double>(x(i,dim)) >= lo && static_cast<double>(x(i,dim)) <= hi) mysend++;
     }
     const int my_store_pos = dev.team_scan(mysend,&nsend());
 
     if (my_store_pos+mysend < maxsendlist) {
     mysend = my_store_pos;
       for (int i=teamstart + dev.team_rank(); i<teamend; i+=dev.team_size()) {
-        if (x(i,dim) >= lo && x(i,dim) <= hi) {
+        if (static_cast<double>(x(i,dim)) >= lo && static_cast<double>(x(i,dim)) <= hi) {
           sendlist(iswap,mysend++) = i;
         }
       }
